create seperate folder for vector_store configs
