+++ docs/components/vectordb.mdx
@@ -51,7 +51,7 @@ from mem0 import Memory
 
 config = {
     "vectordb": {
-        "provider": "chromadb",
+        "provider": "chroma",
         "config": {
             "collection_name": "test",
             "path": "db",
