+++ mem0/configs/vector_stores/chroma.py
@@ -1,0 +1,26 @@
+from typing import Optional,ClassVar
+
+from pydantic import BaseModel, Field, field_validator, model_validator
+
+class ChromaDbConfig(BaseModel):
+    try:
+        from chromadb.api.client import Client
+    except ImportError:
+        raise ImportError("Chromadb requires extra dependencies. Install with `pip install chromadb`") from None
+    Client: ClassVar[type] = Client
+
+    collection_name: str = Field("mem0", description="Default name for the collection")
+    client: Optional[Client] = Field(None, description="Existing ChromaDB client instance")
+    path: Optional[str] = Field(None, description="Path to the database directory")
+    host: Optional[str] = Field(None, description="Database connection remote host")
+    port: Optional[str] = Field(None, description="Database connection remote port")
+
+    @model_validator(mode="before")
+    def check_host_port_or_path(cls, values):
+        host, port, path = values.get("host"), values.get("port"), values.get("path")
+        if not path and not (host and port):
+            raise ValueError("Either 'host' and 'port' or 'path' must be provided.")
+        return values
+    
+    class Config:
+        arbitrary_types_allowed = True
